requirements.txt
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
filelock==3.20.0
fsspec==2025.12.0
google-api-core==2.29.0
google-api-python-client==2.188.0
google-auth==2.47.0
google-auth-httplib2==0.3.0
google-auth-oauthlib==1.2.4
googleapis-common-protos==1.72.0
httplib2==0.31.1
idna==3.11
Jinja2==3.1.6
llvmlite==0.46.0
MarkupSafe==2.1.5
more-itertools==10.8.0
mpmath==1.3.0
networkx==3.6.1
numba==0.63.1
numpy==1.26.4
oauthlib==3.3.1
openai-whisper==20250625
proto-plus==1.27.0
protobuf==6.33.4
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==2.23
pyparsing==3.3.1
regex==2026.1.15
requests==2.32.5
requests-oauthlib==2.0.0
rsa==4.9.1
scipy==1.17.0
sounddevice==0.5.3
sympy==1.14.0
tiktoken==0.12.0
torch==2.2.2+cpu
tqdm==4.67.1
typing_extensions==4.15.0
uritemplate==4.2.0
urllib3==2.6.3
